business logic updates
Tried to move all business logic to the Business project, modified the
JS to re-number the dependent rows when a row is deleted, changed all
doubles to decimals, and a few other small tweaks.

+++ Business/ConfigItems.cs
@@ -19,18 +19,18 @@ namespace Business
         IConfigItemRepository configRepo = new MockConfigItemsRepository();
 
         public int PayPeriods { get; set; }
-        public double YearlyCost { get; set; }
-        public double DependentCost { get; set; }
+        public decimal YearlyCost { get; set; }
+        public decimal DependentCost { get; set; }
         public string AvailableDiscount { get; set; }
-        public double Discount { get; set; }
+        public decimal Discount { get; set; }
 
         public ConfigItems()
         {
             PayPeriods = Int32.Parse(configRepo.GetConfigItem("PayPeriods"));
-            YearlyCost = double.Parse(configRepo.GetConfigItem("YearlyCost"));
-            DependentCost = double.Parse(configRepo.GetConfigItem("DependentCost"));
+            YearlyCost = decimal.Parse(configRepo.GetConfigItem("YearlyCost"));
+            DependentCost = decimal.Parse(configRepo.GetConfigItem("DependentCost"));
             AvailableDiscount = configRepo.GetConfigItem("AvailableDiscount");
-            Discount = double.Parse(configRepo.GetConfigItem("Discount"));
+            Discount = decimal.Parse(configRepo.GetConfigItem("Discount"));
         }
     }
 }
